@@ -1,10 +1,5 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.FileProviders;
-using System;
-using System.Collections.Generic;
+﻿using Microsoft.Extensions.FileProviders;
 using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Microsoft.AspNetCore.Builder
 {
